bcc-22 implemented player, inv and cards

+++ beancards/src/main/java/com/bbdgrads/beancards/repositories/CardRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bbdgrads.beancards.entities.Card;
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
 }
